feat: Add changes to Axios and Fetch Clients interceptors
